(PDB-2482) Add ability to pass in token auth (RBAC auth) to PuppetDB
This commit adds the ability to pass in token authorization functions to
our PuppetDB middleware, these functions come from the `clj-rbac-client`
and will only be configured in the PE PuppetDB extensions repo.

This commit allows us to pass in the `valid-token->subject` and
`is-permitted?` so that our cert-whitelist middleware can optionally
fallback to checking if there is a `X-Authentication` header in the
request and thus a `jwt` with which PuppetDB uses to talk to RBAC. If
the `jwt` has the right permission `puppetdb:use:*` (similar to what the
Orchestrator uses for it's RBAC permission), the request is validated.