(PDB-2482) (PDB-2676) Add ability to use token auth with PuppetDB
This commit adds the clj-rbac-client dependency to PE PuppetDB and uses
the newly added ability from the FOSS middleware to pass in the token
auth functions from the RBAC client.

This commit adds an alias `lein pdb` and some certs/configs to be able
to run PuppetDB, RBAC and Acitivity in the same JVM instance.