(docs) Make the sidebar nav version-agnostic
This commit makes it so we don't have to manually update the version number in
documentation/_puppetdb_nav.html every time we merge master to stable.

Instead of using a hardcoded version, we use existing metadata in the docs site
(configured in puppet-docs/source/_config.yml under the "documents:" key, which
we already update every time we post a new version of the docs). This lets us
look up both the correct version number -- {{ page.doc.version }} -- and the
base URL that the `documentation` folder gets mounted at -- {{puppetdb}}.