(maint) Explicitly configure defaults for web-router-services
This commit explicity configures our defaulted web-router-services to
use the default server using the map syntax so that other services can
run next to PuppetDB in the same JVM instance. Prior to this commit the
status and metrics servers couldn't find a valid server to configure
themselves with when using a "default" server explicitly in the
webserver config.