(PDB-2474) Match AST "in" clause multi/non-multi
When given an AST clause like

  ["from" something ["in" x ["extract" y ["select_facts"]]]]

where one of x and y is a :multi type and the other is not, convert the
:multi's type to its variant that matches the type of the other.

That is, given

  ["from" "facts" ["in" "value" ["extract" "name" ["select_facts"]]]]

convert the fv.value (multi) to fv.value_string to match "name".

To do this, propagate "in" clause columns through the engine as their
{:type ... :field ... ...} maps instead of just as their fields, so that
the type information for both sides of each comparison will be available
once the InExpression node has been assembled.

Adjust the tests to accommodate, add new tests for relevant cases, and
add some schemas to the engine to help document the the compilation
process and expectations.