(maint) Add puppet-agent libs to load path
When we build puppetdb, we require access to the facter library.
Unfortunately, this is under the Ruby Lib path installed with
puppet-agent. System ruby, which we are currently using to build,
doesn't have access to this libdir. This appends the puppet-agent
vendor_lib path to the search path so system ruby can find Facter.