(PDB-2620) Remove runtime validation of reports in storage
We currently validate reports in puppetdb/command.clj and then
re-validate those reports twice in puppetdb/scf/storage.clj. This second
and third validation is very CPU intensive. Some load testing has
revealed as much as 30% of the CPU usage was dedicated to this purpose.

This inefficientcy is mainly spent on unchanged resources and would
likely only be observed on catalogs that have a large number of
unchanged resources.