(PDB-2576) Expire nodes in single SQL command
Instead of expiring one node at at time with a round-trip through PDB
for each, combine stale-nodes and expire-node! into expire-stale-nodes
which handles the entire process in a single SQL command, and returns
all of the expired certnames.

In addition to changing the expiration semantics a bit, this also alters
the reporting, logging "Auto-expired node foo" for each expired node
only after all of the nodes have been expired.  Previously an
"Auto-expiring node foo" was displayed for each node, just before that
node was expired.