(PDB-2650) Fix convergence tests that hit historical catalog limits
This fixes a bug in our tests, when the number of replace-catalog commands
exceeds the historical catalog limit, by simply temporarily replacing
that value with the maximum number of commands generated (unlikely this would
occur, but still entirely possible).

The patch also upgrades test.check so we can pass in a seed via an environment
variable if we want to replicate failures easily next time.

Signed-off-by: Ken Barber <ken@bob.sh>