(PDB-2680) Move PDB status code to a separate ns
This commit moves status related functions to a new status namespace and
factors the integration of TK status into 3 pieces. The `status-details`
which contains the data used to make a decision as to whether PuppetDB
is starting/running/error and a function taht makes that decision. The
third piece is a call to the TK status register function in
pdb-routing. This will allow a different status decision in the PE
version of PuppetDB