(PDB-2704) Log at-at exceptions
The way at-at's uses Java Executors causes any exceptions thrown by a
task to be silently swallowed, so wrap the task with a try/catch that
logs exceptions as errors.

You can demonstrate the problem by adding a throw to the
db-maintenance-tasks function and running this:

  lein test :only puppetlabs.puppetdb.acceptance.node-ttl/test-node-ttl

It should hang.  After this patch, the test should still hang, but
should also log the exception.