(PDB-2737) Added more logging to the threadpool
This commit adds an uncaughtExceptionHandler to the threads used by the
command processing threadpool. This ensures that if for whatever reason
an exception is not caught and handled properly by our command
processing code and it bubbles up to the threadpool, that exception is
logged. This should not happen and we should handle the exceptions
correctly in the command processing code, but this is a safeguard in
case we do not.

This also adds some logging around RejectedExceptionExceptions which can
happen when you are not able to submit work to the threadpool, such as
when it's shutting down.