(PDB-2834) Request :unbuffered-stream, not :stream
When :as :stream is specified, http-client reads the entire request body
into RAM, and then returns a stream wrapping the buffer.  For a request
body over about 2GB, this causes an exception when the underlying
library tries to allocate a byte[] array that's too large (the upper
limit appears to be just under 2GB).

Further, the sync code was written to expect true incremental streaming
of the body.  To arrange that, specify :as :unbuffered-stream to
http-client instead.