(PDB-2847) Add a gated threadpool, fed by channel
This commit adds code that allows data from an core.async input channel
to be processed on a bounded threadpool. When the threadpool reaches
it's max size, it blocks the calling thread on put. The threadpool is
shutdown once the input channel has been closed and in-flight work as
finished.