(PDB-2651) travisci/test.sh: show versions clearly
Print all of the various build versions clearly before running the
tests.

Don't depend on TRAVIS_BRANCH, and --pdb-ref and --[no-]pg-sandbox
options, and pass any options after -- to "lein test", so that it's easy
to run the script anywhere with adequate github access, i.e.

  ext/travisci/test.sh --pdb-ref stable --no-pg-sandbox

Add a --lein argument to replace the use of "$1", and if not specified,
look for a suitable lein/lein2 in the path.