(PDB-2865) Add quick-retries to PDB command processing
Current versions of PuppetDB incur a high cost when a command is
retried. The command gets written to a scheduler queue, then re-enqueued
in the regular queue, then reprocessed. A very common case is a database
transaction issue, where two commands are attempting to update the same
data. One thread wins, one loses and has to be retried. We should
implement a quicker "inner loop" retry, where we don't incur the I/O
overhead in this common case. When the quick retries fail, we can then
go through the normal retry logic.

This commit implements this quick retry in command processing (separate
from the queue logic) and removes the exponential backoff for the
traditional retry logic in favor of a set 10 minute retry time.