(PE-14579) Fix race condition in initial sync
The previous code erroneously assumed that the  pdb threadpool dochan
function blocked until all messages finished processing. Actually, it
only blocked until the channel was empty. This led to a race condition
in which the sync status was updated to be 'idle' before the last
message was actually processed. Upon trying to process the last message,
the status atom was inconsistent with the actual state of the program,
which caused a schema check to fail and an exception to be thrown.

The actual way to finish processing is to shut down the thread pool.
This will block until all pending tasks have been processed.