(ORCH-1491) Better PQL examples and docs structure
This patch adds a new PQL examples page with a list examples containing:

* a title
* description
* pql
* output

It also makes the PQL documentation a top-level element in navigation so it
and other related articles can be found easier.

The tutorial and references pages also have links to each other now so they
are easier to find.

Signed-off-by: Ken Barber <ken@bob.sh>