(PDB-2939) Don't try to store same report twice
There's no reason to, and the attempt causes PDB to wander off into the
long (eventually fruitless) retry process.

We believe that's also what was causing one of the convergence-test
failures.  i.e. if the database connection was slow enough, two syncs
that were scheduled one right after the other could enqueue the same
command twice (in that case the addition of a report), and then the test
would fail (time out) because the second addition never showed up as a
completed command since it was repeatedly rejected as a duplicate.