(PDB-2962) Missing implicit inventory relationships
This adds relationships from environment and node entities to inventory to
behave in a similar way to factsets. This allows implicit subqueries to work.

Without this patch you cannot do queries of this nature:

    nodes { inventory { facts.operatingsystem = "CentOS" } }

Instead you are forced to use explicit querying:

    nodes { certname in inventory[certname] { facts.operatingsystem = "CentOS" } }

Which is a lot more verbose and should be unnecessary.

This also documents the new relationships.

Signed-off-by: Ken Barber <ken@bob.sh>