(maint) Have s3 resolve symlinks for apt repo sync
This commit updates the s3 command used to sync out apt repos to s3.
This change is meant to help enable the switchover from reprepro to
aptly for apt repo creation. In order to maintain the current expected
repo format with aptly, we are forced to create symlinks. However, since
s3 does not support symlinks, we need to ensure they are resolved before
we attempt to send those files to s3. The addition of this flag handles
that.