(PDB-2953) Add code to spot old facts/catalogs before procesing
This commit adds a new core.async channel buffer tracks an
enqueued, but not yet processed factset or catalog for a certname. When
a new catalog/factset is submited, it will update the old
catalog/factset command to be marked for deletion and enqueue the new
catalog normally. When that old message is processed, the only action
taken will be to discard the command (from stockpile).