(PDB-2953) Refactor in-mem commands to cmdrefs
An enqueued message in stockpile contains an id and a metadata
string. This refactor breaks apart the pieces of the metadata string and
stores the items separately when put in the core.async channel. This
allows making changes based on command type in the core.async channel
buffer. That data is then converted into a metadata string when
retrieved from stockpile. This should have a lower memory footprint and
also allow handling of different kinds of messages in different ways in
the core.async channel.