(PDB-2988) Fix pg 9.4.9 concurrent-fact-updates
Fix the concurrent-fact-updates test to work with newer versions of
postgres (9.4.9 at least).

It appears that postgres has changed the way it handles the relevant
transaction, rejecting it for its environments constraint violation
instead of the serialization error.

Establish the relevant "DEV" environment at the start of the test so
that it's no longer part of the transaction, and then change the code to
examine the entire exception chain instead of just the first item
because the serialization exception is no longer at the front.