(PDB-2953) Fix the `remove!` sorted-command-buffer code
An older version of the code used a keyword to reference the "replace
catalog" and "replace facts", that old code didn't get updated to use
the real proper command names. Eventually this will change to an
interned string or a keyword, but this change gets it up and working
with the current command processing code.