(PDB-2905) Migrate dlo to stockpile
Migrate the dlo to stockpile and simplify it.  Stop periodically
archiving messages, and don't pay attention to filesystem changes after
startup.  Perform an initial scan for existing discards at startup, and
only reflect pdb's changes in the metrics afterward.

Store the discards via stockpile's discard command to a file matching
the command's name in the pdb queue, and write information describing
the failure to a corresponding *-err.txt file.  Until we rework the
attempts to be exceptions, just write the strings we have.