(PDB-2734) Add stockpile upgrade code
This commit adds the ability to upgrade from an existing ActiveMQ queue
to Stockpile. This happens automatically on startup. If an ActiveMQ
queue is present, it moves all of the messages out of the queue and into
Stockpile. It leaves a "lock" file in place so that there's no need to
initialize the ActiveMQ broker and connections upon restart.