(PDB-3047) Improve event-counts and aggregate-event-counts endpoints on PE to provide information about corrective changes.
successes and noops counts were subdivided into intentional and corrective parts. So, successes and noops counts are replaced with intentional_successes, corrective_successes, intentional_noops, and corrective_noops. Events are mapped to the corresponding counts based on value of corrective_change flag.

The response formats are unchanged for FOSS users.