(PDB-3039) Escape and truncate certnames as needed
Change `puppetdb.queue/metadata-str` to replace illegal filename
characters and underscores in the certname it's given, truncate the
certname if its length would cause the metadata string to exceed 255
characters, and add the SHA1 hash of the original certname if the
certname in the metadata string differs in any way (via character
replacement or truncation) from the original certname.

Add tests of the new sanitization & truncation behavior of
`puppetdb.queue/metadata-str`.