(MAINT) Uncomment and fix old deactivate node tests
The previous tests included a bare string for the body of the payload in
the command tests. This scenario isn't possible for version 1 or 2 of
the deactive nodes command as our http/command code will always JSON
encode those bare strings. This commit escapes those strings like they
would be if they flowed through from an HTTP request.