(PDB-3039) DLO filenames won't exceed 255 bytes.
Change the *-filename functions in `puppetlabs.command.dlo` to truncate
just the metadata portions of the filenames if they would cause the
filenames to exceed 255 bytes in length. Truncating the metadata portion
is acceptable since all these filenames contain stockpile IDs, which
ensures they'll be unique.

Move the `parse-cmd-filename` function, which has to agree with the DLO
on the format of its discarded message filenames, from puppetdb.queue
over to the puppetdb.command.dlo namespace, to make it easier to see the
need for this concordance and to achieve it.