(PDB-2982) Abbrev command names in queue metadata
Replace `puppetdb.queue/metadata-command-names` with a map called
`metadata-command->puppetdb-comand` that defines the mapping between
the command names that can appear in the queue metadata that forms the
message filenames and the command names used by the rest of PuppetDB.
The inverse mapping is dynamically constructed under the var
`puppetdb-command->metadata-command`.

Replace `puppetdb.queue/metadata-str` with `metadata-serializer`, which
takes a map from PuppetDB command names to the metadata command names
and returns a function with the same signature as `metadata-str`; viz.
taking all the queue metadata components and returning a metadata
string. The only differences are that the returned function looks up the
command name it's given in the translation map and puts the metadata
name it finds there in the metadata string, and that if this lookup
fails then an IllegalArgumentException is thrown.

Bind a metadata serializer with the default translation map of
`queue/puppetdb-command->metadata-command` to `queue/serialize-metadata`

Change `metadata-parser` to take a translation map rather than
a sequential collection comand names, have the returned function perform
the inverse mapping from the metadata command name back to the PuppetDB
command name, and to add a docstring to `metadata-parser`.

Add `parse-metadata` and `serialize-metadata` vars to the
`puppetdb.command.dlo` namespace that add an "unknown" -> "unknown"
entry to both translation maps, since the DLO uses that command name as
a fallback when the message is so garbled that the command name cannot
be recovered. Consequently, change all invocations of
`queue/metadata-str` in the DLO to call `dlo/serialize-metadata`, and
likewise `queue/metadata-parser` to `dlo/parse-metadata`.

Change `puppetdb.command.dlo/discard-cmdref` to create its own queue
metadat string and use `stock/entry` directly in order to create
a stockpile entry for the discarded cmdref, rather than using
`queue/cmdref->entry`, as the latter creates a metadata string using
only the default command names & metadata abbreviations, so it will now
throw an error if the discarded cmdref's command is "unknown".

Update tests to align with the new function names and arguments, as well
as the abbreviated command names in queue metadata strings.