(maint) Add `regex-quote` & `match-any-of` utils.
Add `regex-quote` to the utils NS, which takes a string and returns
a new string that quotes all of the special regex characters in the
first string so it will be interpreted only literally.

Add `match-any-of` which takes a collection of strings and characters,
and returns a regex string with one capturing group that matches any of
the members of the collection literally (using `regex-quote`).