(PDB-2983) Hide rejected execution exceptions.
Because PuppetDB's command-processing threadpools are unbounded, the
only reason that they would throw a RejectedExecutionException is
because they're shutting down or have shut down, which means that
PuppetDB is shutting down. Commands that are rejected are retried when
PuppetDB next starts up, so there's no action the user needs to take in
order to prevent loss of data. Therefore, there's no reason to show the
user the exceptions that arise when the threadpool throws
a RejectedExecutionException. To that end, swallow slingshot exceptions
with kind `:puppetlabs.puppetdb.threadpool/rejected` in the thread
started by `puppetdb.mq-listener/message-listener-service`'s TK start
method.