(PDB-3067) Improve existing command logging.
Change the message identifier in the command namespace's log/info
invocations to be a concatenation of the stockpile ID and the timestamp
the message was received at, in order to guarantee uniqueness across
restarts (as the stockpile ID may reset).

Change the overall structure of the logged message in the same form so
it's clear *what* is being logged.