(PDB-3058) version 6 reports incorrectly rejected
This was a schema error introduced with the corrective_change feature.
The main operational impact was that users upgrading from 3.2.x to 4.2
with enqueued data would have their commands fail on restart.

This commit corrects the schema to accept v6 reports, and also does a
bit of test cleanup so we won't have this problem again. Specifically it
consolidates synthesis of old command versions under the example reports
namespace and causes test compilation to fail if the synthetic reports
are invalid.