(PDB-3207) Add chmod calls to migration and ssl-setup subcommands
In ezbake 1.1.0+, which a previous commit updated the project.clj to,
subcommands run with a umask of 027.  This commit includes a couple of
explicit chmod calls in the migration and ssl-setup subcommands which
ensure that any modified config files are still world-readable.
World-readability is required for these files in order for the
puppetdb service, running as user puppetdb, to be able to still read
the config files, which are owned by root:root.