(PDB-3077) Upgrade to lein-ezbake 1.1.2, fix Debian startup problems
This commit upgrades the lein-ezbake dependency from 1.0.0 to 1.1.2.
This is needed to fix problems with the PuppetDB service not starting
properly on Debian-based platforms due to a permissions issue around the
use of the service rundir.  The rundir problem was previously introduced
by the bump to lein-ezbake 1.0.0.

In the newer ezbake release, subcommands run with a umask of 027.  This
commit includes a couple of explicit chmod calls in the migration and
ssl-setup subcommands which ensure that any modified config files are
still world-readable.  World-readability is required for these files in
order for the puppetdb service, running as user puppetdb, to be able to
still read the config files, which are owned by root:root.