(MAINT) Bump beaker to 2.50.0
Previously, PuppetDB's Gemfile was picking up beaker version 2.43.0.
This was transitively picking up hocon version 0.9.5.  puppetserver
recently added some variable interpolation to the puppetserver.conf
file.  Some fixes in hocon 1.0 and later are needed in order for the
interpolation to be handled correctly.

This commit bumps the beaker dependency in the Gemfile to ~> 2.50.0,
which transitively will cause test jobs to pick up a newer hocon gem
dependency, >= 1.0.