(PDB-3101) Disable certname "bash in place"
This feature has issues in an HA setting. To fix it, we need to use the
producer timestamp to determine which catalog/factset is "newer" before
deciding that we want to discard one of the catalogs/factsets. Disabling
the feature with this commit to be revisited after an initial release.