(MAINT) Fix concurrent command proc test issue
The intention of the tests is to verify that if we receive two factsets
for the same certname and process them at the same time, that we get
correct results (i.e. one wins, the other is retried). These tests were
failing in Jenkins due to a timing issue. Switching from the
SynchronousQueue to a CountDownLatch should fix this issue and avoid
having timeouts like the code did previously.