(PDB-3191) Attempt to flush output before exit
While we may still want to pursue a more coherent exit strategy
eventually (PDB-2053), this hopes to make it less likely that we'll lose
output (diagnostic or otherwise) when we quit via System/exit.

The underlying issue is that the JVM does not necessarily flush stdout
or stderr during a System/exit.