Change where clause on fact nodes to operate on smaller table
Prior to this commit, the where clause operated on the fact_values
table which is a giant table compared to the value_types table.

After this comit, the where clause operates on the value_types
table.  The <> 5 filter forces a seq_scan of the table so doing
that on a table with 6 rows is better than doing it on a table with
100K rows.  Also, filtering on the smaller table first improves the
row estimates in the rest of the query as well.