(docs) Nav: Convert nav sidebar to Markdown with docs site's `{% md %}` tag
Previously, we couldn't use Markdown in the nav sidebar because it was outside
Jekyll's "content" area. But we recently added a
`{% md %} ...stuff... {% endmd %}` tag, which can activate the normal Markdown
converter in places where it's not normally enabled. (For example, outside the
content area, or inside a `<div>` element.)

This means we can simplify the nav snippet and make its tree structure
more legible.