(PDB-3108) Add producer-timestamp as a command header
This commit accepts the producer-timestamp as a POST parameter and
threads that parameter down the stack into the enqueuing code. The
producer-timestamp is required to enable "bash in place". Without it,
we could potentially bash a newer catalogs with a retried (or delayed)
older catalog. Getting the timestamp from the header means we can
continue to delay parsing of the body of the command until it's
dequeued.

It's possible to not have a producer-timestamp for a command. Examples
of that would be on import or if commands were in the queue right
before an upgrade. When there is no producer timestamp present, we
should NOT allow that command to be bashed and instead should process
the command normally (i.e. FIFO). This commit also includes the
necessary logic to consider producer-timestamp when deciding to bash
the command (i.e. mark as deleted, then when dequeued, it's removed).