(PDB-3108) Consolidate producer-timestamps in the terminus
This commit changes the terminus to ensure producer timestamps are
created from a single place. Prior to this patch they were created in
several places, including having the ability to be passed in from
Puppet itself (not possible since producer-timestamp is something PDB
created). This commit also switches to 3 digit accuracy on
timestamps. Prior to this we were using different accuracy for
catalogs/facts vs. reports, both of which were just truncated once it
was received by the PDB backend.