(PDB-3051) benchmark: return stop fn, not channel
Move the --runinterval code to a cycle-commands function that returns
the close channel, but then have benchmark itself wrap that in a stop
function so that the caller can determine whether or not the process is
finished by just testing for a non-nil result.

Change main to sleep forever, making it clearer that any code in main is
irrelevant, since it's not currently coordinated with the shutdown hook.
i.e. on ctrl-c, *no* non-daemon thread code will ever run again, and
daemon thread code will only run (unpredictably) until all finalization
and shutdown hooks have finished.