(PDB-3247) Bump timeout in testutils sync-command-post to 20sec
This commit bumps the testutils sync-command-post to 20sec aligning it
with the 4.x branches of PuppetDB, which should avoid some errors caused
by Jenkins testing performance.

This commit also changes the sync-command-post to not use ex-info and
throw a formated string error instead so that lein test will format the
actual error response instead of hiding it.