(PDB-3108) Refactor command enqueues to use a "command request" map
Prior to this commit, the command enqueuing functions required many
parameters to be passed all the way down the stack. This made adding a
new parameter (in this case producer-timestamp) a tedious
process. This commit switches to a "command request" map and validates
that data in the command enqueuing functions.