(MAINT) Close datasource in max connection jdbc pool tests
Previously, the max connections jdbc pool-construction tests did not
close the connection pool data source before falling out of scope.
Depending upon timing, this could cause the tests to fail on
attempting to drop temp databases which are still busy being used
by the connection pool.

This commit wraps usage of the data source in each of the max
connections tests with `with-open` calls, allowing the connection pool
to be closed off before any calls to drop the temporary databases occur
(and, therefore, for the tests to complete successfully).