(PDB-2640) Remove hardcoding of gzip from http command code
This commit removes the hardcoding of content-encodings to validate
from the http command code, instead relying upon a
`supported-content-encodings` function which derives its list from the
same map used to translate a content-encoding to a file-extension.

This commit also includes an additional test which validates that the
absence of a content-encoding header in the HTTP request is considered
valid.