(PDB-3328) Make run-puppet-as signature closer to run-puppet
It's natural to want to swap back and forth between these, but having
'extra-puppet args' as an optional last argument on one and an options map as an
optional last argument on the other leads to some very strange exceptions. Since
the former is only used once, just make run-puppet-as take an opts map like
run-puppet does.