(PDB-3297) Don't act as if facts might be arrays
In fact, there only a small number of array items in the current
database, yet we check nearly every value that comes back to see if it
might be an array that needs conversion.  Start stopping that.

Accordingly, don't attempt array conversion in call-with-query-rows, and
add a new call-with-array-converted-query-rows that preserves the old
behavior.  Continue to call the former in the fact related storage code,
but adjust everything else to use the latter to preserve existing
behavior.

As the original motivation, Wyatt noticed non-trivial cost for the
"isa?" test via coerce-to-array when processing facts.