(PDB-3420) Exclude inactive and expired nodes in root query endpoint
Previously, queries to the root endpoint (which includes all PQL queries)
included all expired and deactivated nodes. This is different from all the other
query endpoints, and practically never what you want to do.

With this change, the root query endpoint modifies queries in the same way as
the other query endpoints to exclude inactive and expired nodes.