(maint) speed up include_total
include_total is extremely slow on some queries because the current
mechanism does not allow the planner to see through a bunch of json
processing in cases where selected fields include json. This changes the
mechanism of include total from simply wrapping the query sql with a
select count, to modifying the query plan into something more performant
and compiling it separately.