(PDB-3515) Add edges_certname_fkey constraint
We had this before, but it was lost in a migration which rewrote the certnames
table but neglected to restore all of the foreign key constraints. This adds it
back in a special startup function; nominally this would be in a migration, but
we've already consumed the following migration numbers in newer versions. The
corresponding change in the new code will use a migration.