(PDB-3527) Use #to_a method, not #to_data_hash when converting TagSets
The `Puppet::Util::TagSet` class has a method named `#to_data_hash`. It
returns an `Array` instead of a `Hash`. PuppetDB assumes that the hash
returned from a call to `Puppet::Resource::Catalog#to_data_hash` contains
tag sets and calls the `#to_data_hash` to convert them into an `Array`.
This now breaks because the bug that caused the hash to contain tag sets
was fixed in PUP-7381.

This commit changes the call to `#to_data_hash` to instead be a call to
`#to_a`. Both the `Puppet::Util::TagSet` and the `Array` class implements
this method which means that PuppetDB will be capable of munging catalogs
produced both before and after PUP-7381.