(PDB-3544) Improve facts_factset_id_index
Prior to this commit, the facts table had an index on factset_id.

After this commit, we update that index to by on factset_id,
fact_path_id because these two columns are often filtered on together
or one is filtered on while the other is used in a join.