(PDB-3529) Check latest_report_timestamp in one query
Prior to this commit, we found the latest_report_timestamp in a
select query and then compared it to the producer timestamp being
processed in clojure.

After this commit, we just submit an UPDATE query with a
WHERE clause that will only update if the producer timestamp being
processed is newer than the one already stored in the certnames
table.