(PDB-3529) Check latest report on certnames instead of reports
Prior to this commit, we calculated the latest report for a node
by querying the reports table, sorting all the records, and grabbing
the most recent report id.

After this commit, we store the latest report timestamp in the
certnames table with the report-id.  This allows us to compare
each new report with the latest report timestamp as they come in
and decide whether this report is the latest report or not without
querying the reports table.