(PDB-3529) Fix latest-report test and add new test for old reports
Prior to this commit, the latest-report test was comparing
end_time instead of producer timestamp which was incorrect.

After this commit, the latest-report test compares producer timestamp
and has a new test that ensures a report with an older timestamp than
the current latest report will not overwrite the latest report.