(PDB-3541) Don't hang on unsupported database
Instead of calling System/exit, use the trapperkeeper shutdown service
to shut down when an unsupported database is detected.  Previously pdb
would just hang.

One drawback to this approach is that it looks like the only way to
get trapperkeeper to shut down and exit with a non-zero status is to
throw an exception, and we'd prefer to avoid the backtrace since the
problem has already been described in the log and on stderr.

Hopefully trapperkeeper will add something more suitable, and then we
can use it.